Changes necessary for redirect detection
This code effectively checks for redirects while minimizing the number
of http requests necessary to make it happen.

It will ensure that duplicate entries for the same content are not
saved to the DB when the auditor runs simply because of an outdated
link (i.e.: two URLs that link to the same page via permanent redirect).

Note that this does not apply to temporary redirects (302/303/307), but
it could be done easily by adjusting the string parsing in
spider.class.php line 445.

Regarding translations, line 55 of auditorcomponent.php needs better
string management in that this text results in a front-end status
update like this:

‘Redirected to existing URL' crawled successfully! Waiting for crawling
next page for 10 seconds

Which probably should be fixed. I honestly couldn’t figure out where
this display code is managed.